requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
flask-bcrypt==1.0.1
python-dotenv==1.0.0